+++ test/dummy/config/initializers/recording_studio_accessible.rb
@@ -3,6 +3,7 @@
 RecordingStudioAccessible.configure do |config|
   config.access_actor_types = [
     "User",
+    "RecordingStudioApi::ApiClient",
     "RecordingStudioOauth::OauthAuthorization"
   ]
 end